rename pre to loading
