+++ src/teehr/loading/joined_timeseries.py
@@ -5,7 +5,7 @@ from pyspark.sql.types import IntegerType, FloatType
 from pathlib import Path
 from typing import Union
 import teehr.const as const
-from teehr.pre.utils import (
+from teehr.loading.utils import (
     validate_dataset_structure,
 )
 import logging
