+++ src/teehr/loading/locations.py
@@ -1,17 +1,17 @@
 """Module for importing locations from a file."""
 from typing import Union
 from pathlib import Path
-import teehr.pre.utils as tpu
+import teehr.loading.utils as tpu
 import duckdb
-from teehr.pre.duckdb_sql import (
+from teehr.loading.duckdb_sql import (
     create_database_tables,
     insert_locations,
 )
-from teehr.pre.utils import (
+from teehr.loading.utils import (
     validate_dataset_structure,
 )
 from teehr.models.data_tables import locations_field_names
-from teehr.pre.utils import merge_field_mappings
+from teehr.loading.utils import merge_field_mappings
 import teehr.const as const
 
 import logging
